refactor(fleet): extract health/dashboard HTTP server (P4.1)
Move the per-process health/dashboard HTTP server (startHealthServer +
getUiStatus, ~248 lines) and its companion extractWebToken helper out
of FleetManager into fleet-health-server.ts.

- HealthServerContext / UiStatusContext interfaces describe the narrow
  surface FleetManager exposes; FleetManager implements both implicitly
  via existing fields and methods.
- startHealthServer() now returns { server, webToken, startedAt } so the
  caller writes them onto the FleetManager fields it already owns.
- getUiStatus stays on FleetManager as a thin forwarder (required by
  WebApiContext).
- extractWebToken moves alongside the server; the only test importer
  (tests/extract-web-token.test.ts) is updated to point at the new file.

No behaviour change. Avoids a circular import with fleet-manager.ts.

P4.1 step 4 of 4.
